Changes to the graph view basics
